refactor to insert required functions into elements_ephys namespace
